Refactored page component into server and client components
